move cookie code and test
