+++ starsky/starsky/clientapp/src/shared/cookie/get-cookie.ts
@@ -2,5 +2,5 @@ export function GetCookie(name: string): string {
   const regex = new RegExp("(^| )" + name + "=([^;]+)");
   const match = regex.exec(document.cookie);
   if (match) return match[2];
-  return "X-XSRF-TOKEN";
+  return "";
 }
