+++ starsky/starsky/clientapp/src/shared/fetch-post.ts
@@ -1,5 +1,5 @@
 import { IConnectionDefault } from "../interfaces/IConnectionDefault";
-import { GetCookie } from "../get-cookie.ts";
+import { GetCookie } from "./cookie/get-cookie.ts";
 const FetchPost = async (
   url: string,
   body: string | FormData,
